receiver split WORK IN PROGRESS

fixes in e2e test for new receieve architecture - more fixes pending

fixes in receive and e2e test of receive


fixed connection reset by peer error in receive e2e tests. all e2e tests pass


pkg/route: Fix unit tests

route: Make hashring only work on labels

route: Incorporate #2899

fixed import issues and changed the query test signature


made some corrections while rebasing receiver changes


Added a method Error() to implement golang error interface


linting


use the new promauto package


removed unused params


rename method and add comments


removed unused flags in receiver service in e2e tests


replaced promauto with prometheus, as earlier was causing unnecessary panic


changes for debug


removed entry for receive test for e2e


added a readiness probe


fix E2E tests for receive router


revert changes for cmd/main


nitpicks


removed redundant health check


Add receive router registration


not to be commited


modified pkg/receive/handler.go


modified pkg/receive/multitsdb.go


modified pkg/route


modified cmd/thanos/receive_route.go


replaced MultiError with NonNilMultiError


changed to promauto


removed unused variables


changed labelpb.Zlabel to timeseries


whitespace removal and revert the labels


use the main branch's implementation of MultiError
