+++ pkg/route/handler_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) The Thanos Authors.
 // Licensed under the Apache License 2.0.
 
-package receive
+package route
 
 import (
 	"bytes"
@@ -16,16 +16,21 @@ import (
 
 	"github.com/go-kit/kit/log"
 	"github.com/gogo/protobuf/proto"
+	"github.com/gogo/status"
 	"github.com/golang/snappy"
+	"github.com/jpillora/backoff"
 	"github.com/pkg/errors"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	"github.com/prometheus/common/route"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 
 	"github.com/thanos-io/thanos/pkg/errutil"
+	"github.com/thanos-io/thanos/pkg/receive"
 	"github.com/thanos-io/thanos/pkg/store/labelpb"
 	"github.com/thanos-io/thanos/pkg/store/storepb"
 	"github.com/thanos-io/thanos/pkg/store/storepb/prompb"
@@ -175,13 +180,13 @@ func TestDetermineWriteErrorCause(t *testing.T) {
 	}
 }
 
-func newHandlerHashring(appendables []*fakeAppendable, replicationFactor uint64) ([]*Handler, Hashring) {
+func newHandlerHashring(appendables []*receive.FakeAppendable, replicationFactor uint64) (*Handler, []string, Hashring) {
 	cfg := []HashringConfig{
 		{
 			Hashring: "test",
 		},
 	}
-	var handlers []*Handler
+	var endpoints []string
 	// create a fake peer group where we manually fill the cache with fake addresses pointed to our handlers
 	// This removes the network from the tests and creates a more consistent testing harness.
 	peers := &peerGroup{
@@ -197,25 +202,57 @@ func newHandlerHashring(appendables []*fakeAppendable, replicationFactor uint64)
 	}
 
 	for i := range appendables {
-		h := NewHandler(nil, &Options{
-			TenantHeader:      DefaultTenantHeader,
-			ReplicaHeader:     DefaultReplicaHeader,
-			ReplicationFactor: replicationFactor,
-			ForwardTimeout:    5 * time.Second,
-			Writer:            NewWriter(log.NewNopLogger(), newFakeTenantAppendable(appendables[i])),
-		})
-		handlers = append(handlers, h)
-		h.peers = peers
+		h := receive.NewHandler(
+			receive.NewWriter(log.NewNopLogger(), receive.NewFakeTenantAppendable(appendables[i])),
+			nil,
+			DefaultTenant,
+			nil,
+			nil,
+		)
 		addr := randomAddr()
-		h.options.Endpoint = addr
-		cfg[0].Endpoints = append(cfg[0].Endpoints, h.options.Endpoint)
+		cfg[0].Endpoints = append(cfg[0].Endpoints, addr)
+		endpoints = append(endpoints, addr)
 		peers.cache[addr] = &fakeRemoteWriteGRPCServer{h: h}
 	}
 	hashring := newMultiHashring(cfg)
-	for _, h := range handlers {
-		h.Hashring(hashring)
+
+	handler := &Handler{
+		logger: log.NewNopLogger(),
+		router: route.New(),
+		options: &Options{
+			TenantHeader:      DefaultTenantHeader,
+			DefaultTenantID:   DefaultTenant,
+			ReplicationFactor: replicationFactor,
+			ForwardTimeout:    5 * time.Second,
+		},
+		peers: peers,
+		expBackoff: backoff.Backoff{
+			Factor: 2,
+			Min:    100 * time.Millisecond,
+			Max:    30 * time.Second,
+			Jitter: true,
+		},
+		forwardRequests: promauto.With(nil).NewCounterVec(
+			prometheus.CounterOpts{
+				Name: "thanos_receive_forward_requests_total",
+				Help: "The number of forward requests.",
+			}, []string{"result"},
+		),
+		replications: promauto.With(nil).NewCounterVec(
+			prometheus.CounterOpts{
+				Name: "thanos_receive_replications_total",
+				Help: "The number of replication operations done by the receiver. The success of replication is fulfilled when a quorum is met.",
+			}, []string{"result"},
+		),
+		replicationFactor: promauto.With(nil).NewGauge(
+			prometheus.GaugeOpts{
+				Name: "thanos_receive_replication_factor",
+				Help: "The number of times to replicate incoming write requests.",
+			},
+		),
 	}
-	return handlers, hashring
+	handler.Hashring(hashring)
+	return handler, endpoints, hashring
 }
 
 func TestReceiveQuorum(t *testing.T) {
@@ -253,16 +290,16 @@ func TestReceiveQuorum(t *testing.T) {
 		status            int
 		replicationFactor uint64
 		wreq              *prompb.WriteRequest
-		appendables       []*fakeAppendable
+		appendables       []*receive.FakeAppendable
 	}{
 		{
 			name:              "size 1 success",
 			status:            http.StatusOK,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -271,9 +308,9 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -282,9 +319,9 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusConflict,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 			},
 		},
@@ -293,12 +330,12 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -307,15 +344,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -324,15 +361,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -341,15 +378,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -358,15 +395,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -375,18 +412,18 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
-					appenderErr: appenderErrFn,
+					App:         receive.NewFakeAppender(nil, nil, nil, nil),
+					AppenderErr: appenderErrFn,
 				},
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
-					appenderErr: appenderErrFn,
+					App:         receive.NewFakeAppender(nil, nil, nil, nil),
+					AppenderErr: appenderErrFn,
 				},
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
-					appenderErr: appenderErrFn,
+					App:         receive.NewFakeAppender(nil, nil, nil, nil),
+					AppenderErr: appenderErrFn,
 				},
 			},
 		},
@@ -395,15 +432,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusConflict,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 			},
 		},
@@ -412,15 +449,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusConflict,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -429,15 +466,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					App: receive.NewFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -446,15 +483,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -463,15 +500,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusConflict,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					App: receive.NewFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -480,15 +517,15 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					App: receive.NewFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -497,35 +534,36 @@ func TestReceiveQuorum(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
 	} {
 		t.Run(tc.name, func(t *testing.T) {
-			handlers, hashring := newHandlerHashring(tc.appendables, tc.replicationFactor)
+			handler, endpoints, hashring := newHandlerHashring(tc.appendables, tc.replicationFactor)
 			tenant := "test"
-			// Test from the point of view of every node
-			// so that we know status code does not depend
-			// on which node is erroring and which node is receiving.
-			for i, handler := range handlers {
-				// Test that the correct status is returned.
-				rec, err := makeRequest(handler, tenant, tc.wreq)
-				if err != nil {
-					t.Fatalf("handler %d: unexpectedly failed making HTTP request: %v", tc.status, err)
-				}
-				if rec.Code != tc.status {
-					t.Errorf("handler %d: got unexpected HTTP status code: expected %d, got %d; body: %s", i, tc.status, rec.Code, rec.Body.String())
-				}
+			// Test that the correct status is returned.
+			rec, err := makeRequest(handler, tenant, tc.wreq)
+			if err != nil {
+				t.Fatalf("handler %d: unexpectedly failed making HTTP request: %v", tc.status, err)
 			}
+			if rec.Code != tc.status {
+				t.Errorf("got unexpected HTTP status code: expected %d, got %d; body: %s", tc.status, rec.Code, rec.Body.String())
+			}
+
+			// Allow remaining "unnecessary" requests to finish for tests. We
+			// expect them to always succeed in tests unless we explicitly test
+			// error cases.
+			time.Sleep(20 * time.Millisecond)
+
 			// Test that each time series is stored
 			// the correct amount of times in each fake DB.
 			for _, ts := range tc.wreq.Timeseries {
@@ -537,16 +575,14 @@ func TestReceiveQuorum(t *testing.T) {
 					}
 				}
 				for j, a := range tc.appendables {
-					var expectedMin int
-					n := a.appender.(*fakeAppender).Get(lset)
-					got := uint64(len(n))
-					if a.appenderErr == nil && endpointHit(t, hashring, tc.replicationFactor, handlers[j].options.Endpoint, tenant, &ts) {
-						// We have len(handlers) copies of each sample because the test case
-						// is run once for each handler and they all use the same appender.
-						expectedMin = int((tc.replicationFactor/2)+1) * len(ts.Samples)
+					var expected int
+					n := a.App.(*receive.FakeAppender).Get(lset)
+					got := len(n)
+					if a.AppenderErr == nil && endpointHit(t, hashring, tc.replicationFactor, endpoints[j], tenant, &ts) {
+						expected = len(ts.Samples)
 					}
-					if uint64(expectedMin) > got {
-						t.Errorf("handler: %d, labels %q: expected minimum of %d samples, got %d", j, lset.String(), expectedMin, got)
+					if expected > got {
+						t.Errorf("handler: %d, labels %q: expected minimum of %d samples, got %d", j, lset.String(), expected, got)
 					}
 				}
 			}
@@ -589,16 +625,16 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 		status            int
 		replicationFactor uint64
 		wreq              *prompb.WriteRequest
-		appendables       []*fakeAppendable
+		appendables       []*receive.FakeAppendable
 	}{
 		{
 			name:              "size 1 success",
 			status:            http.StatusOK,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -607,9 +643,9 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -618,9 +654,9 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusConflict,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 			},
 		},
@@ -629,12 +665,12 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -643,15 +679,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -660,15 +696,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -677,15 +713,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 1,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -694,15 +730,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -711,18 +747,18 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
-					appenderErr: appenderErrFn,
+					App:         receive.NewFakeAppender(nil, nil, nil, nil),
+					AppenderErr: appenderErrFn,
 				},
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
-					appenderErr: appenderErrFn,
+					App:         receive.NewFakeAppender(nil, nil, nil, nil),
+					AppenderErr: appenderErrFn,
 				},
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
-					appenderErr: appenderErrFn,
+					App:         receive.NewFakeAppender(nil, nil, nil, nil),
+					AppenderErr: appenderErrFn,
 				},
 			},
 		},
@@ -731,15 +767,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusConflict,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 			},
 		},
@@ -748,15 +784,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusConflict,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -765,15 +801,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					App: receive.NewFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -782,15 +818,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusOK,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -799,15 +835,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusConflict,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					App: receive.NewFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					App: receive.NewFakeAppender(conflictErrFn, nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -816,15 +852,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					App: receive.NewFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -833,15 +869,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			status:            http.StatusInternalServerError,
 			replicationFactor: 3,
 			wreq:              wreq1,
-			appendables: []*fakeAppendable{
+			appendables: []*receive.FakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					App: receive.NewFakeAppender(nil, nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					App: receive.NewFakeAppender(nil, nil, nil, nil),
 				},
 			},
 		},
@@ -850,23 +886,21 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 		// to see all requests completing all the time, since we're using local
 		// network we are not expecting anything to go wrong with these.
 		t.Run(tc.name, func(t *testing.T) {
-			handlers, hashring := newHandlerHashring(tc.appendables, tc.replicationFactor)
+			handler, endpoints, hashring := newHandlerHashring(tc.appendables, tc.replicationFactor)
 			tenant := "test"
-			// Test from the point of view of every node
-			// so that we know status code does not depend
-			// on which node is erroring and which node is receiving.
-			for i, handler := range handlers {
-				// Test that the correct status is returned.
-				rec, err := makeRequest(handler, tenant, tc.wreq)
-				if err != nil {
-					t.Fatalf("handler %d: unexpectedly failed making HTTP request: %v", tc.status, err)
-				}
-				if rec.Code != tc.status {
-					t.Errorf("handler %d: got unexpected HTTP status code: expected %d, got %d; body: %s", i, tc.status, rec.Code, rec.Body.String())
-				}
+			// Test that the correct status is returned.
+			rec, err := makeRequest(handler, tenant, tc.wreq)
+			if err != nil {
+				t.Fatalf("handler %d: unexpectedly failed making HTTP request: %v", tc.status, err)
+			}
+			if rec.Code != tc.status {
+				t.Errorf("got unexpected HTTP status code: expected %d, got %d; body: %s", tc.status, rec.Code, rec.Body.String())
 			}
 
-			time.Sleep(50 * time.Millisecond)
+			// Allow remaining "unnecessary" requests to finish for tests. We
+			// expect them to always succeed in tests unless we explicitly test
+			// error cases.
+			time.Sleep(20 * time.Millisecond)
 
 			// Test that each time series is stored
 			// the correct amount of times in each fake DB.
@@ -880,12 +914,12 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 				}
 				for j, a := range tc.appendables {
 					var expected int
-					n := a.appender.(*fakeAppender).Get(lset)
+					n := a.App.(*receive.FakeAppender).Get(lset)
 					got := uint64(len(n))
-					if a.appenderErr == nil && endpointHit(t, hashring, tc.replicationFactor, handlers[j].options.Endpoint, tenant, &ts) {
+					if a.AppenderErr == nil && endpointHit(t, hashring, tc.replicationFactor, endpoints[j], tenant, &ts) {
 						// We have len(handlers) copies of each sample because the test case
 						// is run once for each handler and they all use the same appender.
-						expected = len(handlers) * len(ts.Samples)
+						expected = len(ts.Samples)
 					}
 					if uint64(expected) != got {
 						t.Errorf("handler: %d, labels %q: expected %d samples, got %d", j, lset.String(), expected, got)
@@ -933,13 +967,14 @@ func makeRequest(h *Handler, tenant string, wreq *prompb.WriteRequest) (*httptes
 	if err != nil {
 		return nil, errors.Wrap(err, "marshal request")
 	}
-	req, err := http.NewRequest("POST", h.options.Endpoint, bytes.NewBuffer(snappy.Encode(nil, buf)))
+	req, err := http.NewRequest("POST", "http://example.com/api/v1/receive", bytes.NewBuffer(snappy.Encode(nil, buf)))
 	if err != nil {
 		return nil, errors.Wrap(err, "create request")
 	}
 	req.Header.Add(h.options.TenantHeader, tenant)
 
 	rec := httptest.NewRecorder()
+
 	h.receiveHTTP(rec, req)
 	rec.Flush()
