+++ pkg/route/hashring.go
@@ -1,7 +1,7 @@
 // Copyright (c) The Thanos Authors.
 // Licensed under the Apache License 2.0.
 
-package receive
+package route
 
 import (
 	"context"
@@ -11,7 +11,6 @@ import (
 
 	"github.com/cespare/xxhash"
 	"github.com/pkg/errors"
-
 	"github.com/thanos-io/thanos/pkg/store/storepb/prompb"
 )
 
@@ -112,6 +111,7 @@ func (m *multiHashring) GetN(tenant string, ts *prompb.TimeSeries, n uint64) (st
 	m.mu.RLock()
 	h, ok := m.cache[tenant]
 	m.mu.RUnlock()
+
 	if ok {
 		return h.GetN(tenant, ts, n)
 	}
@@ -159,14 +159,14 @@ func newMultiHashring(cfg []HashringConfig) Hashring {
 	return m
 }
 
-// HashringFromConfigWatcher creates multi-tenant hashrings from a
+// HashringFromConfig creates multi-tenant hashrings from a
 // hashring configuration file watcher.
 // The configuration file is watched for updates.
 // Hashrings are returned on the updates channel.
 // Which hashring to use for a tenant is determined
 // by the tenants field of the hashring configuration.
 // The updates chan is closed before exiting.
-func HashringFromConfigWatcher(ctx context.Context, updates chan<- Hashring, cw *ConfigWatcher) error {
+func HashringFromConfig(ctx context.Context, updates chan<- Hashring, cw *ConfigWatcher) error {
 	defer close(updates)
 	go cw.Run(ctx)
 
@@ -181,19 +181,4 @@ func HashringFromConfigWatcher(ctx context.Context, updates chan<- Hashring, cw 
 			return ctx.Err()
 		}
 	}
-}
-
-// HashringFromConfig loads raw configuration content and returns a Hashring if the given configuration is not valid.
-func HashringFromConfig(content string) (Hashring, error) {
-	config, err := parseConfig([]byte(content))
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to parse configuration")
-	}
-
-	// If hashring is empty, return an error.
-	if len(config) == 0 {
-		return nil, errors.Wrapf(err, "failed to load configuration")
-	}
-
-	return newMultiHashring(config), err
 }
