+++ test/e2e/e2ethanos/services.go
@@ -24,7 +24,7 @@ import (
 	"github.com/thanos-io/thanos/pkg/objstore/client"
 	"github.com/thanos-io/thanos/pkg/query"
 	"github.com/thanos-io/thanos/pkg/queryfrontend"
-	"github.com/thanos-io/thanos/pkg/receive"
+	"github.com/thanos-io/thanos/pkg/route"
 )
 
 const infoLogLevel = "info"
@@ -188,11 +188,7 @@ func NewQuerier(sharedDir, name string, storeAddresses, fileSDStoreAddresses, ru
 
 func RemoteWriteEndpoint(addr string) string { return fmt.Sprintf("http://%s/api/v1/receive", addr) }
 
-func NewReceiver(sharedDir string, networkName string, name string, replicationFactor int, hashring ...receive.HashringConfig) (*Service, error) {
-	localEndpoint := NewService(fmt.Sprintf("receive-%v", name), "", e2e.NewCommand("", ""), nil, 8080, 9091, 8081).GRPCNetworkEndpointFor(networkName)
-	if len(hashring) == 0 {
-		hashring = []receive.HashringConfig{{Endpoints: []string{localEndpoint}}}
-	}
+func NewReceiver(sharedDir string, networkName string, name string) (*Service, error) {
 
 	dir := filepath.Join(sharedDir, "data", "receive", name)
 	dataDir := filepath.Join(dir, "data")
@@ -200,27 +196,19 @@ func NewReceiver(sharedDir string, networkName string, name string, replicationF
 	if err := os.MkdirAll(dataDir, 0777); err != nil {
 		return nil, errors.Wrap(err, "create receive dir")
 	}
-	b, err := json.Marshal(hashring)
-	if err != nil {
-		return nil, errors.Wrapf(err, "generate hashring file: %v", hashring)
-	}
 
 	receiver := NewService(
 		fmt.Sprintf("receive-%v", name),
 		DefaultImage(),
 		// TODO(bwplotka): BuildArgs should be interface.
 		e2e.NewCommand("receive", e2e.BuildArgs(map[string]string{
-			"--debug.name":                 fmt.Sprintf("receive-%v", name),
-			"--grpc-address":               ":9091",
-			"--grpc-grace-period":          "0s",
-			"--http-address":               ":8080",
-			"--remote-write.address":       ":8081",
-			"--label":                      fmt.Sprintf(`receive="%s"`, name),
-			"--tsdb.path":                  filepath.Join(container, "data"),
-			"--log.level":                  infoLogLevel,
-			"--receive.replication-factor": strconv.Itoa(replicationFactor),
-			"--receive.local-endpoint":     localEndpoint,
-			"--receive.hashrings":          string(b),
+			"--debug.name":        fmt.Sprintf("receive-%v", name),
+			"--grpc-address":      ":9091",
+			"--grpc-grace-period": "0s",
+			"--http-address":      ":8080",
+			"--log.level":         infoLogLevel,
+			"--tsdb.path":         filepath.Join(container, "data"),
+			"--label":             fmt.Sprintf(`receive="%s"`, name),
 		})...),
 		e2e.NewHTTPReadinessProbe(8080, "/-/ready", 200, 200),
 		8080,
@@ -233,10 +221,10 @@ func NewReceiver(sharedDir string, networkName string, name string, replicationF
 	return receiver, nil
 }
 
-func NewReceiverWithConfigWatcher(sharedDir string, networkName string, name string, replicationFactor int, hashring ...receive.HashringConfig) (*Service, error) {
+func NewReceiverWithConfigWatcher(sharedDir string, networkName string, name string, hashring ...route.HashringConfig) (*Service, error) {
 	localEndpoint := NewService(fmt.Sprintf("receive-%v", name), "", e2e.NewCommand("", ""), nil, 8080, 9091, 8081).GRPCNetworkEndpointFor(networkName)
 	if len(hashring) == 0 {
-		hashring = []receive.HashringConfig{{Endpoints: []string{localEndpoint}}}
+		hashring = []route.HashringConfig{{Endpoints: []string{localEndpoint}}}
 	}
 
 	dir := filepath.Join(sharedDir, "data", "receive", name)
@@ -259,28 +247,59 @@ func NewReceiverWithConfigWatcher(sharedDir string, networkName string, name str
 		DefaultImage(),
 		// TODO(bwplotka): BuildArgs should be interface.
 		e2e.NewCommand("receive", e2e.BuildArgs(map[string]string{
-			"--debug.name":                              fmt.Sprintf("receive-%v", name),
-			"--grpc-address":                            ":9091",
-			"--grpc-grace-period":                       "0s",
-			"--http-address":                            ":8080",
-			"--remote-write.address":                    ":8081",
-			"--label":                                   fmt.Sprintf(`receive="%s"`, name),
-			"--tsdb.path":                               filepath.Join(container, "data"),
-			"--log.level":                               infoLogLevel,
-			"--receive.replication-factor":              strconv.Itoa(replicationFactor),
-			"--receive.local-endpoint":                  localEndpoint,
-			"--receive.hashrings-file":                  filepath.Join(container, "hashrings.json"),
-			"--receive.hashrings-file-refresh-interval": "5s",
+			"--debug.name":        fmt.Sprintf("receive-%v", name),
+			"--grpc-address":      ":9091",
+			"--grpc-grace-period": "0s",
+			"--http-address":      ":8080",
+			"--log.level":         infoLogLevel,
+			"--tsdb.path":         filepath.Join(container, "data"),
+			"--label":             fmt.Sprintf(`receive="%s"`, name),
 		})...),
 		e2e.NewHTTPReadinessProbe(8080, "/-/ready", 200, 200),
 		8080,
 		9091,
-		8081,
 	)
 	receiver.SetUser(strconv.Itoa(os.Getuid()))
 	receiver.SetBackoff(defaultBackoffConfig)
-
 	return receiver, nil
+}
+
+func NewReceiveRouter(sharedDir string, networkName string, name string, replicationFactor int, hashring ...route.HashringConfig) (*e2e.HTTPService, error) {
+
+	dir := filepath.Join(sharedDir, "data", "receive-route", name)
+	container := filepath.Join(e2e.ContainerSharedDir, "data", "receive-route", name)
+	if err := os.MkdirAll(dir, 0777); err != nil {
+		return nil, errors.Wrap(err, "create router dir")
+	}
+	b, err := json.Marshal(hashring)
+	if err != nil {
+		return nil, errors.Wrapf(err, "generate hashring file: %v", hashring)
+	}
+
+	if err := ioutil.WriteFile(filepath.Join(dir, "hashrings.json"), b, 0666); err != nil {
+		return nil, errors.Wrap(err, "creating router config")
+	}
+
+	router := e2e.NewHTTPService(
+		fmt.Sprintf("receive-route-%v", name),
+		DefaultImage(),
+		// TODO(bwplotka): BuildArgs should be interface.
+		e2e.NewCommand("receive-route", e2e.BuildArgs(map[string]string{
+			"--debug.name":                              fmt.Sprintf("receive-route-%v", name),
+			"--http-address":                            ":8080",
+			"--remote-write.address":                    ":8081",
+			"--log.level":                               infoLogLevel,
+			"--receive.replication-factor":              strconv.Itoa(replicationFactor),
+			"--receive.hashrings-file":                  filepath.Join(container, "hashrings.json"),
+			"--receive.hashrings-file-refresh-interval": "5s",
+		})...),
+		e2e.NewHTTPReadinessProbe(8080, "/-/ready", 200, 200),
+		8080, 8081,
+	)
+	router.SetUser(strconv.Itoa(os.Getuid()))
+	router.SetBackoff(defaultBackoffConfig)
+
+	return router, nil
 }
 
 func NewRuler(sharedDir string, name string, ruleSubDir string, amCfg []alert.AlertmanagerConfig, queryCfg []query.Config) (*Service, error) {
